Authorisation - deleting a pet

+++ src/actions/actions.ts
@@ -99,6 +99,14 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 
 export async function deletePet(petId: unknown) {
   await sleep(1000);
+
+  // authentication check
+  const session = await auth();
+  if (!session?.user) {
+    redirect("/login");
+  }
+
+  //validaation
   const validatedPetId = petIdSchema.safeParse(petId);
   if (!validatedPetId.success) {
     return {
@@ -106,6 +114,25 @@ export async function deletePet(petId: unknown) {
     };
   }
 
+  //authorization check
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validatedPetId.data,
+    },
+  });
+
+  if (!pet) {
+    return {
+      message: "Pet not found.",
+    };
+  }
+
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Not authorised.",
+    };
+  }
+  //database mutation
   try {
     await prisma.pet.delete({
       where: {
